#38: Move integration tests. Tweak docstrings.
